requirements.txt
psycopg2-binary
requests
python-dotenv
pandas
slack-sdk